complete

+++ cell02/ex00/iszero.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-number = float((input()))
+number = int(input())
 if number == 0:
     print("This number is equal to zero.")
 else :
